requirements.txt
click
numpy
Pillow
tensorflow==1.13.1
matplotlib
natsort
tensorflow-hub
imgaug
pypng
pydicom